Move remaining editor tests into tests folder
